==> scripts/01_intro.R <==
source("scripts/viewerpane_html.R")
viewerpane_html("docs/01_intro.html")


## ----eval=FALSE, echo=TRUE---------------------------------------------------------------------------------------------------------------------------------------
## install.packages("readxl")


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
library(readxl)


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
pacman::p_load(readxl,here)


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
?readxl


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
2+2


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
a <- 2+2


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
a


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
a + 4
b <- a+4



## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
b


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
read_excel(here('data','AfricaCovid','AfricaCovid.xlsx'))


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
excel_sheets(here('data','AfricaCovid','AfricaCovid.xlsx'))


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
read_excel(here('data','AfricaCovid','AfricaCovid.xlsx'), sheet="Infected_per_day")


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
africa_covid_cases <- read_excel(here('data','AfricaCovid','AfricaCovid.xlsx'), sheet="Infected_per_day")


## ----eval=FALSE, echo=TRUE---------------------------------------------------------------------------------------------------------------------------------------
## write.csv(africa_covid_cases, here('data', 'africa_covid_cases_long.csv'))


## ----eval=FALSE, echo=TRUE---------------------------------------------------------------------------------------------------------------------------------------
## install.packages("writexl")


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
library(writexl)
write_xlsx(africa_covid_cases, here('data', 'africa_covid_cases.xlsx'))

viewerpane_html('docs/01_intro.html')
