==> scripts/02_data_management.R <==
source("scripts/viewerpane_html.R")
viewerpane_html("docs/02_data_management.html")

## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
name <- "Dave"
number <- 1234
# Check the data type using the function called class
class(name)
class(number)


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
library(readxl)
library(here)
africa_covid_cases <- read_xlsx(here('data', 'africa_covid_cases.xlsx'))
str(africa_covid_cases)


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
mean(africa_covid_cases$`43831`)


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
mean(africa_covid_cases$ISO)


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
africa_covid_cases


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
head(africa_covid_cases, n=5)


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
tail(africa_covid_cases, n=7)


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
unique(africa_covid_cases$COUNTRY_NAME)


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
country_list <- unique(africa_covid_cases$COUNTRY_NAME)


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
unique(africa_covid_cases$AFRICAN_REGION)


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
region_list <- unique(africa_covid_cases$AFRICAN_REGION)


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
tidyverse::tidyverse_packages()


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
head(africa_covid_cases, n=3)


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
library(tidyverse)


## ----wide_to_long------------------------------------------------------------------------------------------------------------------------------------------------
africa_covid_cases_long <-
  africa_covid_cases %>% #tell R to use this dataset
  pivot_longer(cols = 4:492,#select the columns you want
               names_to = "excel_date", #name the new date column
               values_to = "cases") #name the new cases column



## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
head(africa_covid_cases_long, n=3)


## ----piping_example, eval=FALSE, echo=TRUE-----------------------------------------------------------------------------------------------------------------------
## # A fake example of how to bake a cake using piping syntax
## 
## cake <- flour %>%       # to define cake, start with flour, and then...
##   add(eggs) %>%   # add eggs
##   add(oil) %>%    # add oil
##   add(water) %>%  # add water
##   mix_together(         # mix together
##     utensil = spoon,
##     minutes = 2) %>%
##   bake(degrees = 350,   # bake
##        system = "fahrenheit",
##        minutes = 35) %>%
##   let_cool()            # let it cool down


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
#Import population dataset
population_data <- read.csv(here('data','world_population_2021.csv'))
summary(population_data)
head(population_data)


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
names(population_data)


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
names(africa_covid_cases_long)


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
merged_data_left_join <- left_join(africa_covid_cases_long, 
                         population_data,
                         by=c("COUNTRY_NAME"="Location"))
str(merged_data_left_join)


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
merged_data_inner_join <- inner_join(africa_covid_cases_long, 
                         population_data,
                         by=c("COUNTRY_NAME"="Location"))
str(merged_data_inner_join)


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
merged_data_right_join <- right_join(africa_covid_cases_long, 
                         population_data,
                         by=c("COUNTRY_NAME"="Location"))
str(merged_data_right_join)


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
head(africa_covid_cases_long$excel_date)


## ---- eval=FALSE, echo=T-----------------------------------------------------------------------------------------------------------------------------------------
## install.packages("janitor")


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
library(janitor)


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
africa_covid_cases_long <- africa_covid_cases_long %>% 
  mutate(date_format=excel_numeric_to_date(as.numeric(excel_date)))

head(africa_covid_cases_long$date_format)


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
min(africa_covid_cases_long$date_format) #minimum date
max(africa_covid_cases_long$date_format) #maximum date


## ----------------------------------------------------------------------------------------------------------------------------------------------------------------
#remove object called population_data
rm(population_data)
#remove all objects except africa_covid_cases_long and merged_data_right_join
rm(list=ls()[! ls() %in% c("africa_covid_cases_long","merged_data_right_join")])
#remove all objects except africa_covid_cases_long
rm(list=setdiff(ls(), "africa_covid_cases_long"))
write.csv(africa_covid_cases_long, here('data', 'africa_covid_cases_long.csv'))
